refactor: Extract reusable CodeSnippetBlock component
