+++ src/pages/AdvancedSearch/types.ts
@@ -1,0 +1,4 @@
+export enum CodeSnippetView {
+  TREE = "tree",
+  JSON = "json",
+}
